requirements.txt
pandas
numpy
tensorflow
qlib
ib_insync
python-dotenv
requests
